Add explicit newline to get tests to pass
The "passes app FQDN as first argument" test was failing.
It seems that the go-cmd/cmd package doesn't get the final
line of output if there's no newline (it buffers waiting
for newline characters).

Just adding the newline to the test smoketest script for now,
but should probably try to write a failing test for go-cmd/cmd
and submit an issue.